# Add 'Host' header to fix the watcher