@simon3z 2017-05-29
@mindfulmonk can you squash the PR in a single commit? Thanks.