@cben 2017-05-29
@mindfulmonk have you also seen this problem with create/get/update/delete, or only watch?
(I'm asking because these use different HTTP libs.)

I'm not able to reproduce the underlying problem this PR seems to address:
Both with http 0.9.8 and 2.2.2, Host header is set automatically even if I give a headers hash without it:
```
> require 'http'
> puts HTTP.get("https://httpbin.org/headers", headers: {"a" => "b"}).body.to_s
{
  "headers": {
    "A": "b", 
    "Connection": "close", 
    "Host": "httpbin.org", 
    "User-Agent": "http.rb/2.2.2"
  }
}
```
Same with `RestClient.get`.

I'll try to reproduce by actually using kubeclient later.
Can you show the exact URL you pass — is it actually `https://kubernetes/api/`?  Did you by chance use IP address?  Looks so from https://github.com/fabric8io/fluent-plugin-kubernetes_metadata_filter/issues/50, but then I'm not sure how will this PR fix it?  Does a `Host: 192.168.99.100` header actually make Kubernetes happy?