@mindfulmonk 2017-05-30
I have seen this issue only with watch, for example this works fine:

`puts client.get_nodes
`

I have used the fqdn and tested the API with a curl first to ensure it works. 
That being said it fixes the same issue on minishift that I had with using the IP as the fqdn.