@cben 2017-05-30
According to
https://github.com/httprb/http/blob/256539edd9ad8acf0f8a424114ac9fae6fb44782/lib/http/request.rb#L199-L202
http gem also always set a host header (and if we want to set it ourselves
we maybe should include non-default port as that code does?)

Could you repeat the experiment with logging, before and after your patch
(or at least before, I mostly want to understand why it didn't work)?
Add to Gemfile:
    gem "httplog", require: false
and bundle install

Then *after* requiring kubeclient do:
    require 'httplog'
    HttpLog.configure { |config| config.log_headers = true }

then do the watch...