@mindfulmonk 2017-05-30
**Code used:**

```
require 'kubeclient'
require 'rack'
require 'httplog'

HttpLog.configure { |config| config.log_headers = true }

ssl_options = { verify_ssl: OpenSSL::SSL::VERIFY_NONE }
auth_options = {
  bearer_token_file: 'token2'
}
client = Kubeclient::Client.new(
  'https://example-cluster.com/api/', 'v1', auth_options: auth_options, ssl_options: ssl_options
)

watcher = client.watch_pods
watcher.each do |notice|
end
```

**Result without the fix:**

```
D, [2017-05-30T13:38:28.878764 #2114] DEBUG -- : [httplog] Connecting: example-cluster.com:443
D, [2017-05-30T13:38:28.984403 #2114] DEBUG -- : [httplog] Sending: GET http://example-cluster.com:443/api/v1
D, [2017-05-30T13:38:28.984702 #2114] DEBUG -- : [httplog] Header: accept: */*
D, [2017-05-30T13:38:28.984852 #2114] DEBUG -- : [httplog] Header: accept-encoding: gzip, deflate
D, [2017-05-30T13:38:28.984973 #2114] DEBUG -- : [httplog] Header: user-agent: rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
D, [2017-05-30T13:38:28.985078 #2114] DEBUG -- : [httplog] Header: authorization: Bearer xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
D, [2017-05-30T13:38:28.985175 #2114] DEBUG -- : [httplog] Header: host: example-cluster.com
D, [2017-05-30T13:38:28.985426 #2114] DEBUG -- : [httplog] Data: 
D, [2017-05-30T13:38:29.038511 #2114] DEBUG -- : [httplog] Status: 200
D, [2017-05-30T13:38:29.038979 #2114] DEBUG -- : [httplog] Benchmark: 0.034335 seconds
D, [2017-05-30T13:38:29.039213 #2114] DEBUG -- : [httplog] Header: cache-control: no-store
D, [2017-05-30T13:38:29.039360 #2114] DEBUG -- : [httplog] Header: content-type: application/json
D, [2017-05-30T13:38:29.039489 #2114] DEBUG -- : [httplog] Header: date: Tue, 30 May 2017 12:38:27 GMT
D, [2017-05-30T13:38:29.039613 #2114] DEBUG -- : [httplog] Header: transfer-encoding: chunked
D, [2017-05-30T13:38:29.039861 #2114] DEBUG -- : [httplog] Response:
bunch of json here

D, [2017-05-30T13:38:29.049744 #2114] DEBUG -- : [httplog] Sending: GET https://example-cluster.com/api/v1/watch/pods
D, [2017-05-30T13:38:29.049911 #2114] DEBUG -- : [httplog] Header: Authorization: Bearer xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

D, [2017-05-30T13:38:29.049963 #2114] DEBUG -- : [httplog] Header: Connection: close
D, [2017-05-30T13:38:29.049995 #2114] DEBUG -- : [httplog] Header: Host: example-cluster.com
D, [2017-05-30T13:38:29.050024 #2114] DEBUG -- : [httplog] Header: User-Agent: http.rb/0.9.8
D, [2017-05-30T13:38:29.050070 #2114] DEBUG -- : [httplog] Data: 
D, [2017-05-30T13:38:29.050133 #2114] DEBUG -- : [httplog] Connecting: example-cluster.com
D, [2017-05-30T13:38:29.180762 #2114] DEBUG -- : [httplog] Status: 400
D, [2017-05-30T13:38:29.180969 #2114] DEBUG -- : [httplog] Benchmark: 0.130588 seconds
D, [2017-05-30T13:38:29.181151 #2114] DEBUG -- : [httplog] Header: Content-Type: text/plain
D, [2017-05-30T13:38:29.181268 #2114] DEBUG -- : [httplog] Header: Connection: close
D, [2017-05-30T13:38:29.181904 #2114] DEBUG -- : [httplog] Response:
400 Bad Request: missing required Host header
/home/username/.gem/ruby/2.4.1/gems/kubeclient-2.4.0/lib/kubeclient/watch_stream.rb:20:in `each': Bad Request (KubeException)
	from main.rb:16:in `<main>'
```


**Result with the fix**

```
  D, [2017-05-30T13:38:54.801152 #2173] DEBUG -- : [httplog] Connecting: example-cluster.com:443
  D, [2017-05-30T13:38:54.907112 #2173] DEBUG -- : [httplog] Sending: GET http://example-cluster.com:443/api/v1
  D, [2017-05-30T13:38:54.907357 #2173] DEBUG -- : [httplog] Header: accept: */*
  D, [2017-05-30T13:38:54.907464 #2173] DEBUG -- : [httplog] Header: accept-encoding: gzip, deflate
  D, [2017-05-30T13:38:54.907534 #2173] DEBUG -- : [httplog] Header: user-agent: rest-client/2.0.2 (linux-gnu x86_64) ruby/2.4.1p111
  D, [2017-05-30T13:38:54.907592 #2173] DEBUG -- : [httplog] Header: host: example-cluster.com
  D, [2017-05-30T13:38:54.907648 #2173] DEBUG -- : [httplog] Header: authorization: Bearer xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  D, [2017-05-30T13:38:54.907747 #2173] DEBUG -- : [httplog] Data: 
  D, [2017-05-30T13:38:54.958801 #2173] DEBUG -- : [httplog] Status: 200
  D, [2017-05-30T13:38:54.958917 #2173] DEBUG -- : [httplog] Benchmark: 0.038313 seconds
  D, [2017-05-30T13:38:54.958970 #2173] DEBUG -- : [httplog] Header: cache-control: no-store
  D, [2017-05-30T13:38:54.958997 #2173] DEBUG -- : [httplog] Header: content-type: application/json
  D, [2017-05-30T13:38:54.959019 #2173] DEBUG -- : [httplog] Header: date: Tue, 30 May 2017 12:38:53 GMT
  D, [2017-05-30T13:38:54.959040 #2173] DEBUG -- : [httplog] Header: transfer-encoding: chunked
  D, [2017-05-30T13:38:54.959103 #2173] DEBUG -- : [httplog] Response:
bunch of json here

  D, [2017-05-30T13:38:54.962331 #2173] DEBUG -- : [httplog] Sending: GET https://example-cluster.com/api/v1/watch/pods
  D, [2017-05-30T13:38:54.962432 #2173] DEBUG -- : [httplog] Header: Host: example-cluster.com
  D, [2017-05-30T13:38:54.962477 #2173] DEBUG -- : [httplog] Header: Authorization: Bearer xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx

  D, [2017-05-30T13:38:54.962509 #2173] DEBUG -- : [httplog] Header: Connection: close
  D, [2017-05-30T13:38:54.962527 #2173] DEBUG -- : [httplog] Header: User-Agent: http.rb/0.9.8
  D, [2017-05-30T13:38:54.962554 #2173] DEBUG -- : [httplog] Data: 
  D, [2017-05-30T13:38:54.962599 #2173] DEBUG -- : [httplog] Connecting: example-cluster.com
  D, [2017-05-30T13:38:55.103207 #2173] DEBUG -- : [httplog] Status: 200
  D, [2017-05-30T13:38:55.103322 #2173] DEBUG -- : [httplog] Benchmark: 0.140598 seconds
  D, [2017-05-30T13:38:55.103463 #2173] DEBUG -- : [httplog] Header: Cache-Control: no-store
  D, [2017-05-30T13:38:55.103518 #2173] DEBUG -- : [httplog] Header: Content-Type: application/json
  D, [2017-05-30T13:38:55.103549 #2173] DEBUG -- : [httplog] Header: Date: Tue, 30 May 2017 12:38:53 GMT
  D, [2017-05-30T13:38:55.103598 #2173] DEBUG -- : [httplog] Header: Transfer-Encoding: chunked
  D, [2017-05-30T13:38:55.104021 #2173] DEBUG -- : [httplog] Response:

  /home/username/.gem/ruby/2.4.1/gems/http-0.9.8/lib/http/response/body.rb:54:in `stream!': body has already been consumed (HTTP::StateError)
  	from /home/username/.gem/ruby/2.4.1/gems/http-0.9.8/lib/http/response/body.rb:20:in `readpartial'
  	from /home/username/.gem/ruby/2.4.1/gems/http-0.9.8/lib/http/response/body.rb:26:in `each'
  	from /home/username/.gem/ruby/2.4.1/gems/kubeclient-2.4.0/lib/kubeclient/watch_stream.rb:24:in `each'
  	from main.rb:16:in `<main>'


```